requirements.txt
azure-ai-openai
azure-search-documents
requests
ddgs
pathlib
